Fix usage of temporary restricted keep clear area for hotseat
Add proto logs for easier debugging of keep clear area changes.

Originally the change was meant to be harmless, as it would have been
quickly updated with next keepClearAreasChanged callback.
However, we have switched the shelf height to no longer use the API
and instead provide the Rect to avoid directly to PiP, which stopped
issuing the callback and that lingering restricted keep clear area was
hanging over the future invocations for PiP.

It is eventually cleared once the callback is called, but otherwise
creates inconsistent behavior.

Bug: 292813143
Test: manually, enter YT PiP, observe shelf height change movements
Test: after http://recall/-/g8x7ZkgdfbqscttAA9wzl9/f6qcvXMt8A6s5LsC0Rx9Hs
Test: before http://recall/-/g8x7ZkgdfbqscttAA9wzl9/hfcWIBOuMrddsDM5qMv4xj

Change-Id: Ia644bf4ea8c887c7612608bfd867041cb5d8b32a