Reimplement ApkVerityTest
This is a regression/end-to-end test of fs-verity. The original test
setup uses ".fsv_sig" to enable fs-verity on APK then use them as the
target to tamper with.

Now that an app API exists, switch the test setup to use the API. We
can also avoid the placeholder app and signature now.

Since the test is almost a complete rewrite, move it from presubmit to
postsubmit until it's proven stable enough.

Bug: 277916185
Bug: 296465999
Test: atest FsVerityTest
Change-Id: I1bbeb5d6adc960bf3fe4e9c1262ffc7538062f61