Migrate MediaProjectionManager#addCallback to @EnforcePermission
Use the recommended mechanism for permission enforcement in services.
This change is not *technically* behavior preserving because the manual
implementation used checkCallingPermission (as opposed to
checkCalling*orSelf*Permission.

It is safe to make this change since MANAGE_MEDIA_PROJECTION is very
locked down and orSelf simply covers calls within system server.

Bug: 275368789
Test: TH
Change-Id: I603dce46dab1a18c12bc5c25e796bf4e64e46fab