Reland: Deprecate API: isAppSourceCertificateTrusted
Diff from previous patch: changed the flag's package namespace

In the original plan, "app source" such as a store may generate signature
over an app file and install with it. A trusted app source is configurable
by each OEM currently. A next step (that did not happen) is to allow
users to add their trusted app source.

The API is supposed to be used by a store/installer to decide whether
they should download .fsv_sig or not.

As this approach is not scalable, and in fact has not been used at all
per our analysis, we're deprecating the feature, including this API.

When the deprecation feature flag is on, the API simply returns false
(i.e. telling the client no store is trusted and don't bother to
download .fsv_sig; although no one is using it).

A next step is to mark this API @removed in the next release.

Bug: 277916185
Bug: 277348711
Test: flipping the flag android.security.deprecate_fsv_sig
      atest android.security.cts.FileIntegrityManagerTest
Change-Id: I2ec2a0fbcaf902374e2c01d916fed9a8c0bc5a76