Fix tests broken by @NonNull annotation fix
b/294110802 fixed the @NonNull annotation so that it was visible to
kotlin when applied to methods outside the current build module.  This
caused kotlin to insert more null checks, which broke some tests that
use mockito's when/thenReturn syntax, as that calls the method and
mockito may return null where kotlin doesn't expect it.  Fix them by
switching to the doReturn/when syntax, which doesn't call the method.

Bug: 294110802
Bug: 297916459
Bug: 297919489
Test: atest PackageManagerServiceTests FrameworksMockingServicesTests
Change-Id: I0ff1ce3200f73c5596e1f06ea653ff397b9024cb