Make shadeExpansion support all status bar states
Previously, ShadeInteractor.shadeExpansion did not have the correct
value in the SHADE status bar state. Despite the complexity of the
logic that calculates the value, the result is a much simpler
shade expansion value which only provides the amount that the shade
is expanded without any of the edge cases of the legacy codebase.

Test: manually verified by checking all values against go/shade-cujs
Test: added new tests and ran affected tests
Bug: 280887022
Change-Id: I65fe9e0cfef5b3a3ea7b95af6d0039a8f1da1cc0