Prevent screenshot UI from being removed if only one screenshot fails
This changes the logic to report screenshot errors to ScreenshotHelper.

Before this cl, if any display screenshot was failing for any reason, all the UI would have been removed, from all displays, as the TakeScreenshotService was sending a SCREENSHOT_MSG_PROCESS_COMPLETE to ScreenshotHelper that was resetting the connection.

Now, even if one display fails, the SCREENSHOT_MSG_PROCESS_COMPLETE is sent only after the UI of the other display is removed.

Note that this is unrelated with the notification of screenshot error.

Bug: 297355218
Bug: 290910794
Test: TakeScreenshotExecutorTest
Change-Id: I6e6d107048b46200dbff90167f5da37577e72421