Explicitly encode SkImages with Png encoder
Skia plans to stop the default behavior of encoding SkImages in
SkPictures using a PNG encoder. To preserve that behavior,
clients need to explicitly set an SkSerialProc to do so.

The only other call which I could find that made use of the
implicit SkPicture::serialize behavior was Picture [1].
The API is "soft removed" and thus we don't need to maintain
the behavior of serializing or deserialization as long as
it does not crash (which it should not).

[1] frameworks/base/libs/hwui/jni/Picture.cpp

See also: http://review.skia.org/753496

Bug: b/40045064

Change-Id: I2df9bffc5d1046869b5971b0fdd6ff4f1a2216af