Don't consume swipe gesture if screen lock is not secure
If the screen lock is set to swipe or none, swiping up from the dream
will not go to the bouncer, so previously we prevented the animation
from occurring, but by consuming the gesture, it also prevented the
dream from doing anything, meaning that nothing happened when swiping
up.

This change stops consuming the gesture so that the underlying dream
will receive it and can wake the device.

Bug: 298716569
Test: atest BouncerSwipeTouchHandlerTest
Change-Id: If5a495974cfbe5b49433460146116f028fdb4007