Optimize the order of setting remote animating state
When starting recents or back to home, the recents/home is already
the top-app, so setRunningRemoteTransitionDelegate doesn't need to
be called before dispatching animation.

Also if an app is started via core with RemoteTransition set in
ActivityOptions (e.g. launch app from home), then the remote animator
will be set by TransitionController#updateRunningRemoteAnimation.

Because the most common cases already have the top-app/animating state
for the remote animator, it is fine to notify starting animating first
to avoid potential latency of setRunningRemoteTransitionDelegate,
such as scheduling of binder thread and lock contention on core.

Bug: 279437990
Bug: 297493149
Test: Home has top schedule group before starting animation.
      E.g. Launch app from home, swipe to home,
           press home key to home, press back key to home.
Change-Id: I6454e519951eeeda48c7d4a7d582dfa08dec7459