[SB][Scene] Decide when to show user switcher on keyguard in repos.
NPVC has a lot of logic in a lot of different methods about when to show
the user switcher. After reading and re-reading, I think it distills
down to config_keyguardUserSwitcher &&
userManager.isUserSwitcherEnabled(qs_show_user_switcher_for_single_user).
This CL adds a flow to that effect and listens to it.

Bug: 300484307
Test: On phone, verify user switcher is never shown (only user avatar if
there are multiple users, nothing if single user)
Test: On tablet, verify user switcher chip is shown if there are multiple
users (or user & guest), not shown if single user
Test: atest KeyguardStatusBarRepositoryImplTest

Change-Id: I5ec9a8a964219fdea3f7addf95f03469585113f2