Hardcode device aware permissions in system server
CAMERA and MICROPHONE are the only two permissions that can be device aware at this moment. Hardcode them in the system server to
enable saving permission states for device aware permissions. Note: this is a temporary change. Once we have a formal API to return
device aware permissions, we should replace the hardcode.

Bug: 292252664
Test: Existing tests pass. Refer to tests in ag/24855864
Change-Id: If706ea1c9dc5b171a5210298dab5ad4812d6bb88