PSS: Make sure recent task thumbnails are up to date
Task thumbnails are usually only updated when opening them, closing
them, or when going to overview.
This was causing PSS app selector to show an out of date thumbnail
for the last foreground task.
The solution is to make sure to take a new thumbnail when the app
selector starts.

Fixes: 304442979
Test: Manual
Test: MediaProjectionAppSelectorControllerTest
Test: ShellRecentTaskListProviderTest
Test: MediaProjectionAppSelectorScreenshotTest
Change-Id: Id82c0cf47a9be7a610095e6c669fc5aa634d2327